+++ module/documents/chat/chatMessageTorg.js
@@ -15,7 +15,8 @@ export class ChatMessageTorg extends ChatMessage {
       const templateData = { ...this.flags.torgeternity.test }; // make copy
       templateData.isOpen = game.settings.get('torgeternity', 'showCheckDetails') ? "open" : "";
       templateData.ownsActor = fromUuidSync(templateData.actor)?.isOwner;
-      if (templateData.target) templateData.ownsTarget = game.actors.get(templateData.target?.id)?.isOwner;
+      for (const target of templateData.targetAll)
+        target.ownsTarget = fromUuidSync(target.uuid)?.isOwner;
 
       const renderedTemplate = await foundry.applications.handlebars.renderTemplate(this.flags.torgeternity.template, templateData);
       html.querySelector('.message-content').innerHTML = await foundry.applications.ux.TextEditor.enrichHTML(renderedTemplate, { secrets: this.isOwner });